requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argcomplete==2.0.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.12.1
backoff==2.2.1
bcrypt==4.1.2
Beaker==1.12.1
beautifulsoup4==4.12.3
blinker==1.7.0
blivet==3.8.2
blivet-gui==2.5.0
boilerpy3==1.0.7
Brlapi==0.8.5
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
data_handling @ file:///media/singularity/Dokumente/Studentenfutter/CogntiveScience/MasterThesis/Code/master_thesis/data_handling
dataclasses-json==0.6.4
datasets==2.18.0
dbus-python==1.3.2
decorator==5.1.1
deepeval==0.21.17
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnf==4.19.2
docopt==0.6.2
docx2txt==0.8
eduvpn-client==4.2.1
eduvpn-common==1.2.0
elastic-transport==8.13.0
elasticsearch==8.13.0
Events==0.5
execnet==2.0.2
farm-haystack==1.22.1
fastapi==0.110.1
fedora-third-party==0.10
ffmpeg==1.4
file-magic==0.4.0
filelock==3.13.1
Flask==3.0.1
flatbuffers==24.3.25
fros==1.1
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
FVS==0.3.4
google-auth==2.29.0
googleapis-common-protos==1.63.0
gpg==1.20.0
greenlet==3.0.3
grpcio==1.62.1
gssapi==1.7.3
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
humanize==3.13.1
icoextract==0.1.4
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflect==7.0.0
iniconfig==2.0.0
InstructorEmbedding==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.1.0
koji==1.34.0
kubernetes==29.0.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.39
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.39
langtable==0.0.65
lazy-imports==0.3.1
libcomps==0.1.20
libdnf==0.73.1
llama-index==0.10.20
llama-index-agent-openai==0.1.6
llama-index-cli==0.1.11
llama-index-core==0.10.21.post1
llama-index-embeddings-huggingface==0.1.4
llama-index-embeddings-instructor==0.1.3
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.4
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.12
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.11
llama-index-readers-llama-parse==0.1.3
llama-index-retrievers-bm25==0.1.3
llama-index-vector-stores-chroma==0.1.6
llama-index-vector-stores-elasticsearch==0.1.6
llama-parse==0.3.9
llamaindex-py-client==0.1.13
lxml==4.9.3
Mako==1.2.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nftables==0.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.17.1
openai==1.14.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
Paste==3.5.3
pathvalidate==2.5.2
patool==1.12
pdfminer.six==20231228
pdfplumber==0.11.0
pefile==2023.2.7
pexpect==4.8.0
pid==2.2.3
pillow==10.3.0
platformdirs==3.9.1
pluggy==1.4.0
ply==3.11
portalocker==2.8.2
posthog==3.1.0
productmd==1.38
progressbar2==3.53.2
prompthub-py==4.0.0
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pulsar-client==3.4.0
pwquality==1.4.5
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pydantic==2.6.4
pydantic_core==2.16.3
pyenchant==3.2.2
pygit2==1.14.0
Pygments==2.17.2
PyGObject==3.46.0
pykickstart==3.48
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyparted==3.13.0
pypdf==4.1.0
pypdfium2==4.29.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pysbd==0.3.4
PySocks==1.7.1
pytest==8.1.1
pytest-repeat==0.9.3
pytest-xdist==3.5.0
python-augeas==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-meh==0.51
python-utils==3.3.3
pytz==2024.1
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
quantulum3==0.9.0
ragas==0.1.6
rank-bm25==0.2.2
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.13.2
rpm==4.19.1.1
rpmautospec==0.6.3
rpmautospec-core==0.1.4
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sentence-transformers==2.5.1
sentencepiece==0.2.0
sentry-sdk==1.44.1
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==69.2.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.6.0
soupsieve==2.5
SQLAlchemy==2.0.28
sseclient-py==1.8.0
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
systemd-python==235
tabulate==0.9.0
Tempita==0.5.2
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.39.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vkbasalt-cli==3.1.1.post1
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
zombie-imp==0.0.2
